Cast median_value_per_set to float in SQL query
Cast median_value_per_set to float for accurate summation.

+++ models/reporting/salesforce_performance.sql
@@ -53,7 +53,7 @@ WITH office_data as
         SUM(COALESCE(workable_leads,0)) as workable_leads,
         COUNT(DISTINCT CASE WHEN market = '999 - Invalid' THEN lead_id END) as ooa_leads,
         SUM(COALESCE(net_sale_count,0)) as net_sale_count,
-        SUM(COALESCE(median_value_per_set,0)) as set_value
+        SUM(COALESCE(median_value_per_set::float,0)) as set_value
         FROM filetered_data
         GROUP BY 1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22
         {% if not loop.last %}UNION ALL
